Remove redux
At the moment, the app is not complex enough to justify adding redux
and the overhead that comes with it, so I simplified the code
and got rid for `redux`.
