+++ package.json
@@ -15,10 +15,8 @@
     "react-dom": "^16.6.0",
     "react-i18next": "^8.3.7",
     "react-jss": "^8.6.1",
-    "react-redux": "^5.1.0",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.1",
-    "redux": "^4.0.1",
     "typeface-roboto": "^0.0.54"
   },
   "scripts": {
