+++ src/index.js
@@ -1,20 +1,42 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 import "./index.css";
 import "typeface-roboto";
-import "./i18n";
 import * as serviceWorker from "./serviceWorker";
-import store from "./store";
-import { Layout } from "./containers/Layout";
+import { Layout } from "./containers/Layout/";
+import i18n from "i18next";
+import Backend from "i18next-xhr-backend";
+import LanguageDetector from "i18next-browser-languagedetector";
+import { reactI18nextModule } from "react-i18next";
+
+i18n
+  // detect user language
+  // learn more: https://github.com/i18next/i18next-browser-languageDetector
+  .use(LanguageDetector)
+  // load translation using xhr -> see /public/locales
+  // learn more: https://github.com/i18next/i18next-xhr-backend
+  .use(Backend)
+  // pass the i18n instance to react-i18next.
+  .use(reactI18nextModule)
+  // init i18next
+  // for all options read: https://www.i18next.com/overview/configuration-options
+  .init({
+    fallbackLng: "en",
+    debug: true,
+    keySeparator: false, // we do not use keys in form messages.welcome
+    interpolation: {
+      escapeValue: false // not needed for react as it escapes by default
+    },
+    react: {
+      wait: true
+    }
+  });
 
 ReactDOM.render(
-  <Provider store={store}>
-    <BrowserRouter>
-      <Layout />
-    </BrowserRouter>
-  </Provider>,
+  <BrowserRouter>
+    <Layout />
+  </BrowserRouter>,
   document.getElementById("root")
 );
